requirements.txt
langchain~=0.2.14
pinecone-client==2.2.4
python-dotenv==1.0.0
streamlit==1.29.0
langchain-community~=0.2.10
langchain-core~=0.2.34
pymongo~=4.13.2
openai~=1.91.0